refactor(modes): hoist shared tool-batch loop helpers into the SDK
mode-default and mode-goal each carried their own copy of the per-batch tool loop:
executeToolUses (run the batch, synthesizing failed results + an abort on mid-batch
cancel) and emitRequestsAndDetectStuck (the load-bearing stuck-loop orphan-result
fix). A fix to one had to be hand-mirrored to the other — exactly the hazard the
duplicated "See the same fix in mode-tool-use" comments warned about.

Move both into @moxxy/sdk (tool-dispatch.ts), parameterized by a StuckLoopReport for
each mode's wording and goal mode's extra goal_stuck plugin_event. Both modes now
compose them. Pure refactor — event sequences and messages unchanged; verified by the
existing mode-default (9/9) and mode-goal (10/10) suites, including the orphan-pairing
and goal_stuck assertions. Outer loop bodies stay per-mode by design.

Retires TECH_DEBT P2 #6.
